requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.34.19
botocore==1.34.19
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-ip-rotator==1.0.14
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.0.7
webencodings==0.5.1
